+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 - Fixed: remaining template util usages
 - Fixed: error when custom overlay icon anchor position not set
 - Fixed: `map.stylesheet` listener was removed in 2.1.0
+- Fixed: unnecessary container call
 
 ## [2.3.0] - 2021-10-29
 - Changed: use twig support bundle to render map